First officially working version without apparent errors.
The last error was that layer.z_points refers to the z distance to
layer.min_z, not to z=0, but in the partial L-matrices min_z was
subtracted again.
The debugging also revealed that in V1 the partial layers were
calculated "backwards" i.e. for the distance to layer.max_z, which
introduced apparent phase jumps in the wave field. This is now also
fixed.